Move navigation definition to content
