+++ src/utils/pagination.ts
@@ -1,11 +1,11 @@
 import findIndex from 'lodash.findindex';
 
+import navigation from '../../content/navigation';
 import {
   IInternalItem,
   IItem,
   isInternalItem,
-  sidebarItems,
-} from '../sidebarItems';
+} from '../types/Item';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const flattenDeep = (arr: any[]) : any[] => Array.isArray(arr)
@@ -23,7 +23,7 @@ const flattenTree = (tree: IItem[]) : IInternalItem[] => flattenDeep(
     ...(node.secondaryItems ? flattenTree(node?.secondaryItems) : []),
   ])));
 
-const flattenedNav = flattenTree(sidebarItems);
+const flattenedNav = flattenTree(navigation);
 
 const findNodeIndex = (currentPath: string): number => findIndex(
   flattenedNav,
